JSON parser generator: Change number type
Change the type for NUMBER members from int to int32_t. While
all the examples in the json.h header show that you should use
an int in your structs, internally the code uses the size of an
int32_t. On platforms where those are not the same size, this is
bound to create a headache.

See also:
https://github.com/zephyrproject-rtos/zephyr/pull/56346