coap_client: Avoid leaking of socket when bind fails
This change attempts to close the socket when bind fails. Otherwise,
that would never happen, and the system would run out of resources
when retrying repeatedly.

Of course, this call might fail, but there is not much we can do about
that (except possibly rebooting, but maybe you'd want a more elaborate
watchdog infrastructure for that feature anyways).