coap_client: Refactor request tracking
Instead of using Zephyr's slightly awkward pending and reply structs,
I implemented the necessary logic myself by introducing a
request struct with all the necessary information (mostly msg ID and
token). This makes some parts of the code much simpler since we don't
need to allocate 3 things independently of each other, handle errors
and then roll back the complete operation.
At it's current form, it requires a bit more memory since the
actual packet buffer is allocated alongside the request. This is of
course necessary to re-send requests, but it's stupid in cases where
an observation is made, since the request struct has to be kept alive
for the lifetime of the observation.
In practice, a thingsboard client only needs to observe the attributes
and maybe RPC calls, so it won't really matter. If it does, it could
easily be optimized at a later time by separating these two allocs -
but this would make things more complicated again.