Add second test case
Since we have multiple global states, that we don't want to reset
in the test-harness, we need to have a refresh by compiling a
different binary for each test. Ztest assumes, that there is no need
for this, so we need to work with testcase and Kconfig to achieve a
different binary for each test.