Change log message about missing access token
The previous message gave the impression that something was wrong even though
it's a normal state if the device is not yet provisioned.
Also clarify *where* the access token is missing because this message
will also be displayed if an access token was given via KConfig
option but it hasn't been written to non volatile storage yet.