Handle non-piggybacked responses
So far the response handler only handled piggybacked responses but
apparently for some confirmable requests thingsboard sends the ACK
and the response separately.

Note that we also don't rely on the ACK when we get a non-piggybacked
response first. This is finde according to RFC 7252:

'Note that, as the underlying datagram transport may not be
sequence-preserving, the Confirmable message carrying the response
may actually arrive before or after the Acknowledgement message for the
request; for the purposes of terminating the retransmission sequence,
this also serves as an acknowledgement.'