Fix scheduling of receive worker
client_cycle_requests reschedules the coap worker. Afterwards the worker is scheduled again
with a delay of APP_RECEIVE_INTERVAL which did not have any effect because k_work_schedule()
instead of k_work_reschedule was used.
The result was that after sending a request the worker did not run for a few seconds
and therefore received the response too late, leading to a retransmission of the request.