Make sure attribute parser code is generated before the app
When building an app that uses the generated attribute parser from
the thingsboard SDK it happened that the code generation ran after
the compile step of the app which failed when including the attribute
parser header.
Creating a custom CMake target and adding it as dependency to the SDK
ensures that code generation runs first.